Split GUI native file drop actions
